pkg/storage: Update or delete more tests related to profile tree deletion
