+++ pkg/storage/series_test.go
@@ -234,269 +234,6 @@ func TestMemSeriesMany(t *testing.T) {
 	}
 }
 
-func TestMemSeriesTree(t *testing.T) {
-	var (
-		label    = map[string][]string{"foo": {"bar", "baz"}}
-		numLabel = map[string][]int64{"foo": {1, 2}}
-		numUnit  = map[string][]string{"foo": {"bytes", "objects"}}
-	)
-
-	// Note: These keys are not unique per location.
-	// For this test they simply seem to be.
-	k2 := ProfileTreeValueNodeKey{location: "00000000-0000-0000-0000-000000000002|00000000-0000-0000-0000-000000000001|00000000-0000-0000-0000-000000000000"}
-	k3 := ProfileTreeValueNodeKey{location: "00000000-0000-0000-0000-000000000003|00000000-0000-0000-0000-000000000001|00000000-0000-0000-0000-000000000000"}
-	k4 := ProfileTreeValueNodeKey{location: "00000000-0000-0000-0000-000000000004|00000000-0000-0000-0000-000000000001|00000000-0000-0000-0000-000000000000", labels: `"foo"["bar" "baz"]`, numlabels: `"foo"[1 2][6279746573 6f626a65637473]`}
-	k5 := ProfileTreeValueNodeKey{location: "00000000-0000-0000-0000-000000000005|00000000-0000-0000-0000-000000000002|00000000-0000-0000-0000-000000000001|00000000-0000-0000-0000-000000000000"}
-
-	s11 := makeSample(1, []uuid.UUID{
-		uuid.MustParse("00000000-0000-0000-0000-000000000002"),
-		uuid.MustParse("00000000-0000-0000-0000-000000000001"),
-	})
-
-	s12 := makeSample(2, []uuid.UUID{
-		uuid.MustParse("00000000-0000-0000-0000-000000000004"),
-		uuid.MustParse("00000000-0000-0000-0000-000000000001"),
-	})
-	s12.Label = label
-	s12.NumLabel = numLabel
-	s12.NumUnit = numUnit
-
-	s := NewMemSeries(0, labels.FromStrings("a", "b"), func(int64) {}, newHeadChunkPool())
-	s.timestamps = append(s.timestamps, &timestampChunk{chunk: chunkenc.FromValuesDelta(1)})
-
-	pt1 := NewProfileTree()
-	pt1.Insert(s11)
-	pt1.Insert(s12)
-	err := s.seriesTree.Insert(0, pt1)
-	require.NoError(t, err)
-
-	require.Len(t, s.flatValues, 2)
-	require.Equal(t, chunkenc.FromValuesXOR(1), s.flatValues[k2][0])
-	require.Equal(t, chunkenc.FromValuesXOR(2), s.flatValues[k4][0])
-
-	require.Len(t, s.labels, 1)
-	require.Equal(t, map[ProfileTreeValueNodeKey]map[string][]string{k4: label}, s.labels)
-	require.Equal(t, map[ProfileTreeValueNodeKey]map[string][]int64{k4: numLabel}, s.numLabels)
-	require.Equal(t, map[ProfileTreeValueNodeKey]map[string][]string{k4: numUnit}, s.numUnits)
-
-	require.Equal(t, &MemSeriesTree{
-		s: s,
-		Roots: &MemSeriesTreeNode{
-			LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000000"), // root
-			Children: []*MemSeriesTreeNode{{
-				LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000001"),
-				Children: []*MemSeriesTreeNode{{
-					keys:       []ProfileTreeValueNodeKey{k2},
-					LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000002"),
-				}, {
-					keys:       []ProfileTreeValueNodeKey{k4},
-					LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000004"),
-				}},
-			}},
-		},
-	}, s.seriesTree)
-
-	// Merging another profileTree onto the existing one
-
-	s2 := makeSample(3, []uuid.UUID{
-		uuid.MustParse("00000000-0000-0000-0000-000000000002"),
-		uuid.MustParse("00000000-0000-0000-0000-000000000001"),
-	})
-
-	pt2 := NewProfileTree()
-	pt2.Insert(s2)
-
-	s.timestamps[0] = &timestampChunk{chunk: chunkenc.FromValuesDelta(1, 2)}
-	err = s.seriesTree.Insert(1, pt2)
-	require.NoError(t, err)
-
-	require.Len(t, s.flatValues, 2)
-	require.Equal(t, chunkenc.FromValuesXOR(1, 3), s.flatValues[k2][0])
-	require.Equal(t, chunkenc.FromValuesXOR(2), s.flatValues[k4][0]) // sparse - nothing added
-
-	// The tree itself didn't change by adding more values but no new locations.
-	require.Equal(t, &MemSeriesTree{
-		s: s,
-		Roots: &MemSeriesTreeNode{
-			LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000000"), // root
-			Children: []*MemSeriesTreeNode{{
-				LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000001"),
-				Children: []*MemSeriesTreeNode{{
-					keys:       []ProfileTreeValueNodeKey{k2},
-					LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000002"),
-				}, {
-					keys:       []ProfileTreeValueNodeKey{k4},
-					LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000004"),
-				}},
-			}},
-		},
-	}, s.seriesTree)
-
-	// Merging another profileTree onto the existing one with one new Location
-	s3 := makeSample(4, []uuid.UUID{
-		uuid.MustParse("00000000-0000-0000-0000-000000000003"),
-		uuid.MustParse("00000000-0000-0000-0000-000000000001"),
-	})
-
-	pt3 := NewProfileTree()
-	pt3.Insert(s3)
-
-	s.timestamps[0] = &timestampChunk{chunk: chunkenc.FromValuesDelta(1, 2, 3)}
-	err = s.seriesTree.Insert(2, pt3)
-	require.NoError(t, err)
-
-	require.Len(t, s.flatValues, 3)
-	require.Equal(t, chunkenc.FromValuesXOR(1, 3), s.flatValues[k2][0])   // sparse - nothing added
-	require.Equal(t, chunkenc.FromValuesXOR(2), s.flatValues[k4][0])      // sparse - nothing added
-	require.Equal(t, chunkenc.FromValuesXORAt(2, 4), s.flatValues[k3][0]) // new
-
-	// The tree itself didn't change by adding more values but no new locations.
-	require.Equal(t, &MemSeriesTree{
-		s: s,
-		Roots: &MemSeriesTreeNode{
-			LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000000"), // root
-			Children: []*MemSeriesTreeNode{{
-				LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000001"),
-				Children: []*MemSeriesTreeNode{{
-					keys:       []ProfileTreeValueNodeKey{k2},
-					LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000002"),
-				}, {
-					keys:       []ProfileTreeValueNodeKey{k3},
-					LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000003"),
-				}, {
-					keys:       []ProfileTreeValueNodeKey{k4},
-					LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000004"),
-				}},
-			}},
-		},
-	}, s.seriesTree)
-
-	// Merging another profileTree onto the existing one with one new Location
-	s4 := makeSample(6, []uuid.UUID{
-		uuid.MustParse("00000000-0000-0000-0000-000000000005"),
-		uuid.MustParse("00000000-0000-0000-0000-000000000002"),
-		uuid.MustParse("00000000-0000-0000-0000-000000000001"),
-	})
-	pt4 := NewProfileTree()
-	pt4.Insert(s4)
-
-	s.timestamps[0] = &timestampChunk{chunk: chunkenc.FromValuesDelta(1, 2, 3, 4)}
-	err = s.seriesTree.Insert(3, pt4)
-	require.NoError(t, err)
-
-	// Merging another profileTree onto the existing one with one new Location
-	s5 := makeSample(7, []uuid.UUID{
-		uuid.MustParse("00000000-0000-0000-0000-000000000002"),
-		uuid.MustParse("00000000-0000-0000-0000-000000000001"),
-	})
-	pt5 := NewProfileTree()
-	pt5.Insert(s5)
-
-	s.timestamps[0] = &timestampChunk{chunk: chunkenc.FromValuesDelta(1, 2, 3, 4, 5)}
-	err = s.seriesTree.Insert(4, pt5)
-	require.NoError(t, err)
-
-	require.Len(t, s.flatValues, 4)
-	require.Equal(t, chunkenc.FromValuesXOR(1, 3, 0, 0, 7), s.flatValues[k2][0])
-	require.Equal(t, chunkenc.FromValuesXOR(2), s.flatValues[k4][0])
-	require.Equal(t, chunkenc.FromValuesXORAt(2, 4), s.flatValues[k3][0])
-	require.Equal(t, chunkenc.FromValuesXORAt(3, 6), s.flatValues[k5][0])
-
-	// The tree itself didn't change by adding more values but no new locations.
-	require.Equal(t, &MemSeriesTree{
-		s: s,
-		Roots: &MemSeriesTreeNode{
-			LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000000"), // root
-			Children: []*MemSeriesTreeNode{{
-				LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000001"),
-				Children: []*MemSeriesTreeNode{{
-					keys:       []ProfileTreeValueNodeKey{k2},
-					LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000002"),
-					Children: []*MemSeriesTreeNode{{
-						keys:       []ProfileTreeValueNodeKey{k5},
-						LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000005"),
-					}},
-				}, {
-					keys:       []ProfileTreeValueNodeKey{k3},
-					LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000003"),
-				}, {
-					keys:       []ProfileTreeValueNodeKey{k4},
-					LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000004"),
-				}},
-			}},
-		},
-	}, s.seriesTree)
-}
-
-func TestMemSeriesTreeMany(t *testing.T) {
-	snano := time.Second.Nanoseconds()
-
-	s := NewMemSeries(0, labels.FromStrings("a", "b"), func(int64) {}, newHeadChunkPool())
-
-	app, err := s.Appender()
-	require.NoError(t, err)
-
-	for i := 1; i < 200; i++ {
-		pt1 := NewProfileTree()
-		pt1.Insert(makeSample(int64(i), []uuid.UUID{
-			uuid.MustParse("00000000-0000-0000-0000-000000000002"),
-			uuid.MustParse("00000000-0000-0000-0000-000000000001"),
-		}))
-		pt1.Insert(makeSample(2*int64(i), []uuid.UUID{
-			uuid.MustParse("00000000-0000-0000-0000-000000000004"),
-			uuid.MustParse("00000000-0000-0000-0000-000000000001"),
-		}))
-
-		err = app.Append(context.Background(), &Profile{
-			Meta: InstantProfileMeta{
-				Timestamp: int64(i),
-				Duration:  snano,
-				Period:    snano,
-			},
-			Tree: pt1,
-		})
-		require.NoError(t, err)
-	}
-
-	it := NewMultiChunkIterator(s.root)
-	for i := 1; i < 200; i++ {
-		require.True(t, it.Next())
-		require.Equal(t, int64(2*i+i), it.At())
-	}
-
-	require.Len(t, s.flatValues, 2)
-
-	it = NewMultiChunkIterator(s.flatValues[ProfileTreeValueNodeKey{location: "00000000-0000-0000-0000-000000000002|00000000-0000-0000-0000-000000000001|00000000-0000-0000-0000-000000000000"}])
-	for i := 1; i < 200; i++ {
-		require.True(t, it.Next())
-		require.Equal(t, int64(i), it.At())
-	}
-
-	it = NewMultiChunkIterator(s.flatValues[ProfileTreeValueNodeKey{location: "00000000-0000-0000-0000-000000000004|00000000-0000-0000-0000-000000000001|00000000-0000-0000-0000-000000000000"}])
-	for i := 1; i < 200; i++ {
-		require.True(t, it.Next())
-		require.Equal(t, 2*int64(i), it.At())
-	}
-
-	// The tree itself didn't change by adding more values but no new locations.
-	require.Equal(t, &MemSeriesTree{
-		s: s,
-		Roots: &MemSeriesTreeNode{
-			LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000000"), // root
-			Children: []*MemSeriesTreeNode{{
-				LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000001"),
-				Children: []*MemSeriesTreeNode{{
-					keys:       []ProfileTreeValueNodeKey{{location: "00000000-0000-0000-0000-000000000002|00000000-0000-0000-0000-000000000001|00000000-0000-0000-0000-000000000000"}},
-					LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000002"),
-				}, {
-					keys:       []ProfileTreeValueNodeKey{{location: "00000000-0000-0000-0000-000000000004|00000000-0000-0000-0000-000000000001|00000000-0000-0000-0000-000000000000"}},
-					LocationID: uuid.MustParse("00000000-0000-0000-0000-000000000004"),
-				}},
-			}},
-		},
-	}, s.seriesTree)
-}
-
 func TestMemSeries_truncateChunksBefore(t *testing.T) {
 	testcases := []struct {
 		before int64
